requirements.txt
openai
python-dotenv
pinecone-client
pandas
fastapi
anthropic